% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{add_margins}
\alias{add_margins}
\title{Adjust ggplot2 margins}
\usage{
add_margins(t = 0.1, r = 0.5, b = 0.1, l = 0.1, unit = "cm")
}
\arguments{
\item{t}{number with top margin}

\item{r}{number with right margin}

\item{b}{number with bottom margin}

\item{l}{number with left margin}

\item{unit}{string with the unit to use. "cm"  by default}
}
\description{
Adjust ggplot2 margins
}
