% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{colores_em}
\alias{colores_em}
\title{Macroeconomic Surveys colors}
\usage{
colores_em(which = NULL)
}
\arguments{
\item{which}{color or palette to pick from the list}
}
\value{
a vector with one or more colors
}
\description{
Macroeconomic Surveys colors
}
