% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{create_month_dirs}
\alias{create_month_dirs}
\title{Crear el directorio para alojar los resultados del procesamiento de la encuesta}
\usage{
create_month_dirs(encuesta, year, mes, path = ".")
}
\arguments{
\item{encuesta}{encuesta a procesar "eem" o "eoe"}

\item{year}{year value}

\item{mes}{string con el mes}

\item{path}{string con la ruta}
}
\description{
Crear el directorio para alojar los resultados del procesamiento de la encuesta
}
