% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{date_label}
\alias{date_label}
\title{Create date labels out of dates}
\usage{
date_label(date = Sys.Date(), year_in_new_line = FALSE)
}
\arguments{
\item{date}{a date vector}

\item{year_in_new_line}{should the year be in it own line?}
}
\value{
a character string
}
\description{
Take a date object into a 'Month yyyymm' format
}
