% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_plots.R
\name{eem_boxplot}
\alias{eem_boxplot}
\title{Create boxplot}
\usage{
eem_boxplot(
  data_eem,
  variable,
  ...,
  min_date = "2022-01-01",
  color = colores_em("blue"),
  font_size = 10,
  labsx = NULL,
  labsy = NULL,
  font_family = "Gotham Book"
)
}
\arguments{
\item{data_eem}{eem data in wide format, or simmilar strutured data frame}

\item{variable}{variable key: "inflacion", "tc", "tcd", "pib" or "tpm"}

\item{...}{ggplot2 theme configuration}

\item{min_date}{start date of the plot in yyyy-mm-dd format}

\item{color}{color to use from the colores_em()}

\item{font_size}{number with the funt size}

\item{labsx}{labs for x axis}

\item{labsy}{labs for y axis}

\item{font_family}{font family name. Use extrafont::load_font() to load}
}
\value{
a ggplot2 object
}
\description{
Designed to plot boxplot out of the wide EEM data. Just provide the data frame
and the name of the variable to plot.
}
\examples{
eem_boxplot(example_eem_data(), "inflacion_interanual", font_family = "sans")
eem_boxplot(example_eem_data(), "tcd_diciembre", font_family = "sans")
}
