% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data.R
\name{eem_data_diciembre}
\alias{eem_data_diciembre}
\title{Forecast for the end of the year}
\usage{
eem_data_diciembre(data_eem, variable = "inflacion", year_to_plot = 2023)
}
\arguments{
\item{data_eem}{eem data in wide format}

\item{variable}{variable key: "inflacion", "tc", "tcd", "pib" or "tpm"}

\item{year_to_plot}{year to plot}
}
\value{
a data frame
}
\description{
In the survey forecast for the end of a given year are collected up to 24 times.
12 times the year before, as end of next year forecasts, and 12 times during the
current year ad en of year expectation. This functions gather those forecasts
into one series
}
\examples{
eem_data_diciembre(example_eem_data(), "inflacion", 2023)
}
