% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_handling_data.R
\name{eem_month_data}
\alias{eem_month_data}
\title{Read and prepare the monthly data}
\usage{
eem_month_data(data_year, data_mes)
}
\arguments{
\item{data_year}{surevey year}

\item{data_mes}{survey month as number}
}
\value{
a data frame
}
\description{
Read and prepare the monthly data
}
