% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_plots.R
\name{eem_ribbon_plot}
\alias{eem_ribbon_plot}
\title{EEM ribbon plot}
\usage{
eem_ribbon_plot(
  eem_long,
  plot_var,
  ...,
  horizon = "12 meses",
  stat = "promedio",
  start_year = 2020,
  ribbon_min = 0.25,
  ribbon_max = 0.75,
  ribbon_alpha = 0.3,
  ribbon_fill = colores_em("gray"),
  color = NULL,
  dot_size = 2,
  breaks = 6,
  font_size = 13,
  font_family = "Gotham Book"
)
}
\arguments{
\item{eem_long}{eem data in long format. Get it using \code{get_eem_data("long")}}

\item{plot_var}{variable to plot: "inflacion", "tc", "tcd", "pib" or "tpm"}

\item{...}{ggplot2 theme configuration}

\item{horizon}{horizon to plot}

\item{stat}{string indicating "promedio" or "mediana". Promedio by default}

\item{start_year}{When should the plot start}

\item{ribbon_min}{a number between 0 and 1}

\item{ribbon_max}{a number between 0 and 1}

\item{ribbon_alpha}{a number between 0 and 1}

\item{ribbon_fill}{color to fill the ribbon. Literal color or hex code}

\item{color}{color for the line. Literal color or hex code}

\item{dot_size}{number indicating the size of the dots}

\item{breaks}{integer with the number of breaks}

\item{font_size}{number with the font size}

\item{font_family}{font family name. Use extrafont::load_font() to load}
}
\value{
a ggplot2 object
}
\description{
Plot median or mean with distribution area. Useful to explore the evolution
of the expectations and it's distribution for a given variable.
}
