% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data.R
\name{example_eem_data}
\alias{example_eem_data}
\title{Get EEM example data}
\usage{
example_eem_data(format = "wide")
}
\arguments{
\item{format}{wide or long format}
}
\value{
a data frame
}
\description{
Get EEM example data
}
\examples{
example_eem_data("wide")
example_eem_data("long")
}
