% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_data.R
\name{get_data_eem}
\alias{get_data_eem}
\title{Returns the historical series of the Macroeconomic Expectation Survey}
\usage{
get_data_eem(format = c("wide", "long"), data_path = NULL)
}
\arguments{
\item{format}{a string indicating the data frame format "wide" or "long"}

\item{data_path}{an optional argument, needed when your are using the package
outside the project it was build to work on}
}
\value{
a data frame with the historical data
}
\description{
Returns the historical series of the Macroeconomic Expectation Survey
}
