% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_observada.R
\name{get_data_obervada_eem}
\alias{get_data_obervada_eem}
\title{Data observada EEM}
\usage{
get_data_obervada_eem()
}
\description{
Descagar y preparar las series observadas de las variables de la EEM
en los distintos horizonte
}
