% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{my_date_breaks}
\alias{my_date_breaks}
\title{Escala de fechas mateniendo fija la fecha más reciente}
\usage{
my_date_breaks(min_date, max_date, step = 4)
}
\arguments{
\item{min_date}{fecha más lejana}

\item{max_date}{fecha más reciente}

\item{step}{distancia entre cada fecha, en meses}
}
\value{
un vector tipo fecha
}
\description{
Esta función es útil para aquellos casos gráficos en los que se requiere
que la fecha más reciente esté visible en el eje.
}
\examples{
my_date_breaks(as.Date("2019-01-01"), Sys.Date())
}
