% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\name{read_details_file}
\alias{read_details_file}
\title{Read details form yaml file}
\usage{
read_details_file(group)
}
\arguments{
\item{group}{string wiht the component needed: "een", "eoe"}
}
\value{
a list with the existing values
}
\description{
Read details form yaml file
}
