% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector given desired values}
\usage{
rescale(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{vector to rescale}

\item{new_min}{new minimum value}

\item{new_max}{new maximum value}
}
\value{
A numeric vector
}
\description{
Rescale a vector given desired values
}
\examples{
rescale(1:10, 0, 1)
}
