% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitarios.R
\name{saving_plot}
\alias{saving_plot}
\title{Guardar gráficos de ggplot2}
\usage{
saving_plot(plot, w, h, path = "./", dpi = 350)
}
\arguments{
\item{plot}{a ggplot2 object}

\item{w}{width in inches}

\item{h}{height in inches}

\item{path}{where to save the plot}

\item{dpi}{dots per inch}
}
\description{
Recibe gráficos y los guarda en PNG usando como bombre el nombre del objeto
}
