% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{theme_em}
\alias{theme_em}
\title{Theme for Macroeconomic Surveys plots}
\usage{
theme_em(
  ...,
  text_size = 11,
  title_size = 16,
  strip_size = 10,
  legend_size = 10,
  legend_position = "bottom",
  font_family = "Gotham Book"
)
}
\arguments{
\item{...}{ggplot2:theme arguments}

\item{text_size}{plot text size}

\item{title_size}{title text size}

\item{strip_size}{strip text size}

\item{legend_size}{legend text size}

\item{legend_position}{legend position}

\item{font_family}{font family name. Use extrafont::load_font() to load
fonts}
}
\value{
a theme object
}
\description{
Theme for Macroeconomic Surveys plots
}
