% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_handling_data.R
\name{update_eem_historic}
\alias{update_eem_historic}
\title{Update the historical data}
\usage{
update_eem_historic(month_data, new_data = TRUE)
}
\arguments{
\item{month_data}{monthly data}

\item{new_data}{is this a new entry?}
}
\description{
Update the historical data
}
